refactor: extract chat thread lifecycle
